Replaced the use of IntVar based entry widgets, excepting Checkbuttons and Radiobuttons, with StringVar based widgets that return either str, int, or Decimal values.  This is to eliminate odd behaver of IntVar, including interpreting values starting with 0 as octal.
Added print option to the SaveBiometricsReportFrame GUI.  The report document is opened via the webbrowser library, so the user can use the broswer's print capability.  This is a temporary solution until a multi-platform print dialog can be developed.

Completely recoded the ViewEditBiometrics GUI to use the same structure as the ScheduleEntryWindow for better UI